client: support external pebble-challtestsrv.
Now that Pebble comes with `pebble-challtestsrv` it's nice to have
`acmeshell` support using its own internal challtestsrv or an external
`pebble-challtestrsv`. This branch accomplishes that with a quick and
dirty interface that translates local challtestsrv functions into POSTs
to the `pebble-challtestsrv` API. The code is gross but so is 90% of
`acmeshell`, sorry!

A few other misc tweaks slipped their way in to this branch. Its a hacky
repo where I do bad things quickly.