Improve newOrder to parse and print all fields
This change adds three new fields to the internal representation of
orders (`expires`, `notBefore`, and `notAfter`), for full compliance
with the set of fields defined by RFC 8555 7.1.3:
  https://tools.ietf.org/html/rfc8555#section-7.1.3

In addition, it adds logic to print the full result of creating a
new order, rather than only printing the URL of the order object.
Finally, it updates the documentation of the `client.CreateOrder`
method to clarify that the given order object is modified in place.